requirements.txt
anthropic>=0.25.0
streamlit>=1.35.0
pdfplumber>=0.10.0
python-docx>=1.1.0